+++ CMakeLists.txt
@@ -143,6 +143,7 @@ set(libsrc
     src/plugins_types_identityref.c
     src/plugins_types_instanceid.c
     src/plugins_types_integer.c
+    src/plugins_types_leafref.c
     src/plugins_types_string.c
     src/plugins_exts.c
     src/plugins_exts_metadata.c
